login page styling, mobile nav menu function
