+++ onbrand-client/package.json
@@ -7,7 +7,7 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "bootstrap": "^5.0.1",
+    "bootstrap": "^5.1.0",
     "graphql": "^15.5.0",
     "graphql-tag": "^2.12.4",
     "jwt-decode": "^3.1.2",
@@ -16,6 +16,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "reactstrap": "^8.9.0",
+    "sass": "^1.37.5",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
