+++ onbrand-client/src/App.css
@@ -18,35 +18,6 @@
     justify-content: space-around;
 }
 
-.form-group {
-    margin: 6px;
-    width: 300px;
-}
-
-.login {
-    background-color: #294f6b;
-    height: 100vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.onbrandlogo {
-    max-height: 84px;
-    max-width: 204px;
-}
-
-.onBrandLogin {
-    max-height: 173px;
-    max-width: 420px;
-}
-
-.login-btn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
 .form-box {
     background-color: #c4c4c4;
 }
