+++ onbrand-client/src/App.js
@@ -4,11 +4,11 @@ import { AuthProvider, AuthContext } from './context/auth';
 import { AuthRoute } from './util/AuthRoute';
 import './App.css';
 
-import Home from './pages/Home';
-import LandingPage from './pages/LandingPage';
-import NavBar from './components/NavBar';
-import Login from './pages/Login';
-import Signup from './pages/Signup';
+import Home from './pages/Home/Home';
+import LandingPage from './pages/LandingPage/LandingPage';
+import NavBar from './components/NavBar/NavBar';
+import Login from './pages/Login/Login';
+import Signup from './pages/Signup/Signup';
 
 function App() {
   const { user } = useContext(AuthContext);
