+++ onbrand-client/src/components/NavBar/NavBar.styles.scss
@@ -1,0 +1,97 @@
+.color-nav {
+  background-color: #294F6B;
+  -webkit-box-shadow: 0px 1px 1px #d4af37;
+  -moz-box-shadow: 0px 1px 1px #d4af37;
+  box-shadow: 0px 1px 1px #d4af37;
+}
+
+.login {
+  background-color: #294f6b;
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.onbrandlogo {
+  max-height: 84px;
+  max-width: 204px;
+}
+
+.onBrandLogin {
+  max-height: 173px;
+  max-width: 420px;
+}
+
+// .navbar-toggler {
+//   padding: 0.5rem !important;
+//   position: relative !important;
+//   background-color: transparent !important;
+//   border-color: transparent !important;
+// }
+
+// .navbar-toggler:focus,
+// .navbar-toggler:active,
+// .navbar-toggler-icon:focus {
+//   outline: none;
+//   box-shadow: none;
+// }
+
+// .navbar-toggler span {
+//   display: block !important;
+//   background-color: #d4af37 !important;
+//   height: 2px !important;
+//   width: 27px !important;
+//   margin-top: 5px !important;
+//   margin-bottom: 5px !important;
+//   transform: rotate(0deg) !important;
+//   left: 0 !important;
+//   opacity: 1 !important;
+// }
+
+// .navbar-toggler:focus,
+// .navbar-toggler:active {
+//   outline: 0 !important;
+// }
+
+// .navbar-toggler span:nth-child(1),
+// .navbar-toggler span:nth-child(3) {
+//   transition: transform 0.35s ease-in-out !important;
+//   transition: transform 0.35s ease-in-out !important;
+// }
+
+// .navbar-toggler:not(.collapsed) span:nth-child(1) {
+//   position: absolute !important;
+//   right: 10px !important;
+//   top: 10px !important;
+//   transform: rotate(135deg) !important;
+//   opacity: 0.9 !important;
+// }
+
+// .navbar-toggler:not(.collapsed) span:nth-child(2) {
+//   height: 12px !important;
+//   visibility: hidden !important;
+//   background-color: transparent !important;
+// }
+
+// .navbar-toggler:not(.collapsed) span:nth-child(3) {
+//   position: absolute !important;
+//   right: 12px !important;
+//   top: 10px !important;
+//   transform: rotate(-135deg) !important;
+//   opacity: 0.9 !important;
+// }
+
+// .navbar-toggler:not(.collapsed) span:nth-child(2) {
+//   height: 12px !important;
+//   visibility: hidden !important;
+//   background-color: transparent !important;
+// }
+
+// .navbar-toggler:not(.collapsed) span:nth-child(3) {
+//   position: absolute !important;
+//   right: 12px !important;
+//   top: 10px !important;
+//   transform: rotate(-135deg) !important;
+//   opacity: 0.9 !important;
+// }
